requirements.txt
# Core
flask>=3.0
python-dotenv>=1.0
openai>=1.0
Pillow>=10.0
pyodbc>=5.0

# Optional: for RAW image support (NEF, CR2, ARW, etc.)
# rawpy>=0.20
# imageio>=2.30
